to module

+++ pose_scene_picker/diff_frame.py
@@ -1,54 +1,60 @@
-
-import sys
 from pathlib import Path
 
 import cv2
 import numpy as np
 
-out_dir = Path(sys.argv[2])
-out_dir.mkdir(parents=True, exist_ok=True)
 
+def pick_pose_frame(
+        movie_file,
+        output_dir,
+        binalization_threshold=30,
+        next_frame_pixel_threshold=20_000):
+    # th = 30    # 差分画像の閾値
 
-th = 30    # 差分画像の閾値
+    output_dir = Path(output_dir)
+    output_dir.mkdir(parents=True, exist_ok=True)
 
-# 動画ファイルのキャプチャ
-cap = cv2.VideoCapture(sys.argv[1])
+    # 動画ファイルのキャプチャ
+    if not Path(movie_file).exists():
+        raise FileNotFoundError(f'{movie_file}')
 
-# 最初のフレームを背景画像に設定
-ret, frame = cap.read()
+    cap = cv2.VideoCapture(movie_file)
 
-# グレースケール変換
-bg = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+    # 最初のフレームを背景画像に設定
+    ret, frame = cap.read()
 
-i = 0
-filepath = out_dir / f"{str(i).zfill(4)}.jpg"
-cv2.imwrite(str(filepath), frame)
-i += 1
+    # グレースケール変換
+    bg = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 
-try:
-    while(cap.isOpened()):
-        # フレームの取得
-        ret, frame = cap.read()
+    i = 0
+    filepath = output_dir / f"{str(i).zfill(4)}.jpg"
+    cv2.imwrite(str(filepath), frame)
+    i += 1
 
-        # グレースケール変換
-        gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+    try:
+        while(cap.isOpened()):
+            # フレームの取得
+            ret, frame = cap.read()
 
-        # 差分の絶対値を計算
-        mask = cv2.absdiff(gray, bg)
+            # グレースケール変換
+            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 
-        # 差分画像を二値化してマスク画像(モノクロ)を算出
-        mask[mask < th] = 0
-        mask[mask >= th] = 255
+            # 差分の絶対値を計算
+            mask = cv2.absdiff(gray, bg)
 
-        print(i, np.count_nonzero(mask))
-        num_of_diff_pixel = np.count_nonzero(mask)
+            # 差分画像を二値化してマスク画像(モノクロ)を算出
+            mask[mask < binalization_threshold] = 0
+            mask[mask >= binalization_threshold] = 255
 
-        if num_of_diff_pixel > 20_000:
-            filepath = out_dir / f"{str(i).zfill(4)}.jpg"
-            cv2.imwrite(str(filepath), frame)
-            cap.read()  # なんかゴミが出てるので飛ばす
-            i += 1
+            print(i, np.count_nonzero(mask))
+            num_of_diff_pixel = np.count_nonzero(mask)
 
-        bg = gray
-finally:
-    cap.release()
+            if num_of_diff_pixel > next_frame_pixel_threshold:
+                filepath = output_dir / f"{str(i).zfill(4)}.jpg"
+                cv2.imwrite(str(filepath), frame)
+                cap.read()  # なんかゴミが出てるので飛ばす
+                i += 1
+
+            bg = gray
+    finally:
+        cap.release()
